Don't return error AND send message on open fail
Since open already returns an error synchronously, it's unnecessary for
it to send a message about the failure as well. This only was done in
active mode and it was an unintentional side effect of always checking
whether file descriptors needed to be added to the poll call.

This also simplifies the unit tests verifying this behavior and makes
them less picky on the particular error returned from open since that
varies between Linux and Windows.