Add documentation for Line framing separator
When working with a LTE modem I needed to specify the framing separator
for the `Circuits.UART.Framing.Line` framer. It took some work to figure
out how to accomplish this with the current documentation.

This tries to provide an example on how to set the separator so if
others run into this it will be more clear for them.